+++ dsptoolbox/__init__.py
@@ -14,6 +14,7 @@ from . import filterbanks
 from . import special
 from . import audio_io
 from . import beamforming
+from . import effects
 
 __all__ = [
     # Basic classes
@@ -28,7 +29,7 @@ __all__ = [
 
     # Modules
     'transfer_functions', 'distances', 'room_acoustics', 'plots', 'generators',
-    'filterbanks', 'special', 'audio_io', 'beamforming'
+    'filterbanks', 'special', 'audio_io', 'beamforming', 'effects'
 ]
 
-__version__ = '0.2.5'
+__version__ = '0.2.6'